+++ CHANGELOG.md
@@ -18,6 +18,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Security
 - in case of vulnerabilities
 
+## [0.0.17] - 2020-01-13
+### Fixed
+- Fixed bug in human readable output of command ltacmd request new
+- Picker will quarantine a TransferRequest if it cannot make a bundle
+
 ## [0.0.16] - 2020-01-06
 ### Changed
 - Command: ltacmd request status now shows information about bundles
@@ -113,7 +118,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Project setup scripts
 
-[Unreleased]: https://github.com/WIPACrepo/lta/compare/v0.0.16...HEAD
+[Unreleased]: https://github.com/WIPACrepo/lta/compare/v0.0.17...HEAD
+[0.0.17]: https://github.com/WIPACrepo/lta/compare/v0.0.16...v0.0.17
 [0.0.16]: https://github.com/WIPACrepo/lta/compare/v0.0.15...v0.0.16
 [0.0.15]: https://github.com/WIPACrepo/lta/compare/v0.0.14...v0.0.15
 [0.0.14]: https://github.com/WIPACrepo/lta/compare/v0.0.13...v0.0.14